feat(web): enhance task management components and introduce new dropdowns
- Refactored `TaskSecondaryInfo` to replace Listbox with DropdownMenu for improved UI/UX.
- Added new components: `ActiveTaskLabelsDropdown`, `ActiveTaskProjectDropdown`, and `MultipleStatusDropdown` for better task status management.
- Introduced hooks for task labels, priorities, sizes, and versions to streamline data handling.
- Improved task status dropdown functionality and integrated new utility functions for mapping task statuses.
- Cleaned up and optimized existing code for better maintainability and performance.
