+++ apps/web/core/hooks/tasks/use-task-sizes-value.ts
@@ -1,0 +1,8 @@
+import { TTaskStatus } from '@/core/types/schemas';
+import { useMapToTaskStatusValues } from './use-map-to-task-status-values';
+import { useTaskSizes } from './use-task-sizes';
+
+export function useTaskSizesValue() {
+	const { taskSizes } = useTaskSizes();
+	return useMapToTaskStatusValues(taskSizes as TTaskStatus[], false);
+}
