+++ apps/web/core/hooks/tasks/use-task-versions-value.ts
@@ -1,0 +1,8 @@
+import { useMapToTaskStatusValues } from './use-map-to-task-status-values';
+import { useTaskVersion } from './use-task-version';
+
+export function useTaskVersionsValue() {
+	const { taskVersion } = useTaskVersion();
+
+	return useMapToTaskStatusValues(taskVersion, false);
+}
